add type hints, move articles to utils, use single logger
