+++ listeners/__init__.py
@@ -1,5 +1,8 @@
+from slack_bolt import App
+
 from listeners import steps
+from utils.news_fetcher import NewsFetcher
 
 
-def register_listeners(app, news_fetcher):
+def register_listeners(app: App, news_fetcher: NewsFetcher):
     steps.register(app, news_fetcher)
