+++ listeners/steps/__init__.py
@@ -1,8 +1,8 @@
 from slack_bolt import App
 
 from listeners.steps import workflow_step
+from utils.news_fetcher import NewsFetcher
 
 
-def register(app: App, news_fetcher):
-    # Your app will answer with the following callback functions when presented with workflow events
+def register(app: App, news_fetcher: NewsFetcher):
     workflow_step.enable_workflow_step(app, news_fetcher)
